requirements.txt
beautifulsoup4==4.11.1
Jinja2==3.1.2
jsonschema==4.16.0
Markdown==3.4.1
PyYAML==6.0
